+++ trunk/Robot/Sonic/System/RunableModule.h
@@ -8,7 +8,7 @@ public:
     RunableModule(void);
     ~RunableModule(void);
     //the main entry point for work each cycle
-    virtual bool RunTick(float time,RobotState state);
+    virtual bool RunTick(uint16_t time,RobotState state);
     //Shows us some debug
     virtual void DebugOutput(void);
 };
